Move some things around and reorganize using autoloader.
